separated web3 and pages logic
