+++ public/js/Toasts.js
@@ -1,12 +1,4 @@
-// const toastTrigger = document.getElementById("liveToastBtn");
 const toastAlert = document.getElementById("liveToast");
-// if (toastTrigger) {
-//   toastTrigger.addEventListener("click", () => {
-//     const toast = new bootstrap.Toast(toastAlert);
-
-//     toast.show();
-//   });
-// }
 
 function CreateToast(danger, msg){
   const toastMsg = document.getElementById("toast-msg");
