+++ views/layouts/main.hbs
@@ -69,6 +69,13 @@
                   for="seller-role"
                 >Seller</label>
               </div>
+              <button
+               id="connectButton" 
+               type="button" 
+               class="btn btn-primary" 
+               disabled
+               >Connected</button>
+
               <div class="col-xl-6 div-address"><p id="address">My Address:
                   <span id="show-account"></span></p></div>
 
@@ -225,8 +232,5 @@
     <script src="js/form-validation.js"></script>
   </body>
 
-  <script src="/js/SQLRequests.js" type="module"></script>
-  <script src="/js/Toasts.js" type="module"></script>
-  <script src="/js/AdditionalLogic.js"></script>
-  <script src="/js/EscrowChain.js" type="module"></script>
+  <script src="js/mainPage.js" type="module"></script>
 </html>
