+++ views/layouts/profile.hbs
@@ -1,0 +1,109 @@
+<script src="js/ethers-5.2.umd.min.js" type="application/javascript"></script>
+<script src="/js/EscrowProvider.js"></script>
+
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="description" content="" />
+    <meta
+      name="author"
+      content="Mark Otto, Jacob Thornton, and Bootstrap contributors"
+    />
+    <meta name="generator" content="Hugo 0.101.0" />
+    <title>Profile</title>
+
+    <link
+      rel="canonical"
+      href="https://getbootstrap.com/docs/5.2/examples/checkout/"
+    />
+
+    <link
+      href="css/bootstrap.min.css"
+      rel="stylesheet"
+      crossorigin="anonymous"
+    />
+    <link rel="stylesheet" href="/css/style.css" />
+
+  </head>
+  <body class="bg-light">
+
+    <div class="base-container container">
+      <main>
+        <div class="row g-5 deal-details">
+          <div class="col-4">
+            <div class="card">
+            <img src="https://avatars.githubusercontent.com/u/78862872?v=4" class="card-img-top" alt="...">
+            
+                <div class="card-body border-top color-border-muted pt-3 mt-3 d-none d-md-block">
+                    
+                    <h5 class="card-title">AveZorgen</h5>
+                    <h6 class="card-subtitle mb-2 text-muted">Rating 4.5/5</h6>
+                    <p class="card-text">Some info from user</p>
+                    <a href="#" class="btn btn-primary">Contact</a>
+                </div>
+            </div>
+          </div>
+          
+          <div class="col-8">
+            <button
+              id="connectButton" 
+              type="button" 
+              class="btn btn-primary" 
+              disabled
+              >Connected</button>
+            <h4 class="d-flex justify-content-between align-items-center mb-3">
+              <span class="text-primary">History</span>
+            </h4>
+
+            <ul class="list-group mb-3" id="history-list">
+              <li class="list-group-item d-flex justify-content-between lh-sm">
+                <div>
+                  <h6
+                    class="my-0"
+                  >0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266</h6>
+                  <small class="text-muted">pending</small>
+                </div>
+                <span class="text-muted">0.1 ETH</span>
+              </li>
+              <li
+                class="list-group-item d-flex justify-content-between bg-light"
+              >
+                <div class="text-success">
+                  <h6
+                    class="my-0"
+                  >0x70997970C51812dc3A010C7d01b50e0d17dc79C8</h6>
+                  <small class="text-muted">1 day ago</small>
+                </div>
+                <span class="text-success">0.005 ETH</span>
+              </li>
+              <li class="list-group-item d-flex justify-content-between">
+                <span>Total (ETH)</span>
+                <strong>0.105 ETH</strong>
+              </li>
+            </ul>
+
+            <form class="card p-2" id="tmp">
+              <div class="input-group">
+                <input type="text" class="form-control" placeholder="Address" />
+                <button type="submit" class="btn btn-secondary">Filter</button>
+              </div>
+            </form>
+
+            <button
+              id="answerCreate-btn"
+              class="w-100 btn btn-primary btn-lg"
+            >Update history</button>
+            <div class="col-xl-6">
+              <p>Massage answer: <span id="answerCreate"></span></p><br />
+            </div>
+          </div>
+        </div>
+      </main>
+    </div>
+
+    <script src="js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
+  </body>
+
+  <script src="/js/profilePage.js" type="module"></script>
+</html>
